// src/app/content-item/latest-content-validation-error-code.enum.ts
export enum LATEST_CONTENT_VALIDATION_ERROR_CODE {
  EMPTY_SOURCES = 'EMPTY_SOURCES',
  INVALID_LIMIT_PER_SOURCE = 'INVALID_LIMIT_PER_SOURCE',
  UNSUPPORTED_SOURCE_KIND = 'UNSUPPORTED_SOURCE_KIND',
  UNSUPPORTED_JOURNAL_API_PROVIDER = 'UNSUPPORTED_JOURNAL_API_PROVIDER',
  INVALID_SOURCE_REQUEST = 'INVALID_SOURCE_REQUEST',
}